refactor: centralize notifications in core dispatch/completion functions (#150)
